Moving map functions to map.go
